enhance: add new list layouts for practice quizzes
